[deadcode]/[fix] `app/assets/stylesheets/application.css`
This file can cause a `Sprockets::DoubleLinkError` in circumstances that, while latent in the state of this commit, are likely to appear in normal Ruby development. Here is how to reproduce the issue:

1. Include some CSS from `application.js`, e.g.
   1. Create an empty `test.css` file in the same directory
   1. Add the following line to `application.js`: <pre>import "./test.css";</pre>
1. Make sure to include `application.css` from a Rails view. For example, if you are still using the stock `vendor/bundle/ruby/*/gems/railties-*/lib/rails/templates/rails/welcome/index.html.erb` view, add this line in its `<head>` section: <pre><%= stylesheet_link_tag "application", "data-turbo-track": "reload" %></pre>
1. Make sure to invalidate your build cache (`rm app/assets/builds/*.*`) and restart Rails (`yarn build; ./bin/rails server -p 3000`)

The problem is that both `app/assets/stylesheets/application.css` and esbuild (the latter, only if using the “import CSS into JavaScript” feature) want to input `application.css` into the asset pipeline. The solution is to

- delete `app/assets/stylesheets/application.css`; developers who want an “old-style” CSS can just create one using another name.
- delete its reference in `app/assets/config/manifest.js` as well