[wip] starting with stimulus for edit page
But, I would like to first try the new _standard_ way of ruby
of serving js. Therefore I park this and create a new branch
where I will try to _back to standard_. Will need to check how
to make Dominique's authentication stuff to work. But, may be
if I stick to standard rails, I won't need it anymore.
We'll see.