[feat] minimal authorisation + custom error pages
Opted for gem action_policy which claims to be cleaner than the good old CanCan and also pundit.
TODO: now, for uniformity, we should use it for the visibility part too.
Error pages are now in their controller but the reason behind this is very weak and I
will just keep it like this for easier editing. Once we will be happy with the result
I think we will go back to the standard static html pages in the /public folder.