[chore] cleanup unused tests
Why removing files that will have to be added later ?
In fact there is a motivation: ./bin/rails test:models is very often failing with an error about
database conneciton being. Sometimes it even freezes in an apparent deadlock. I was hoping
that by reducing the number of tests that could run in parallel I might go around the issue.
Off course this is not the case.
Tests on single files do work (e.g. `./bin/rails test tests/model/award.rb`) therefore
it is really an issue related to parallel test execution.